Move clear_disk_cache to memory::fadvise
