+++ lib/common/memory/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod chunked_utils;
+pub mod fadvise;
 pub mod madvise;
 pub mod mmap_ops;
 pub mod mmap_type;
