+++ lib/gridstore/src/page.rs
@@ -1,7 +1,8 @@
 use std::path::{Path, PathBuf};
 
 use memmap2::MmapMut;
-use memory::madvise::{Advice, AdviceSetting, Madviseable, clear_disk_cache};
+use memory::fadvise::clear_disk_cache;
+use memory::madvise::{Advice, AdviceSetting, Madviseable};
 use memory::mmap_ops::{create_and_ensure_length, open_write_mmap};
 
 use crate::tracker::BlockOffset;
