+++ lib/segment/src/index/field_index/full_text_index/mmap_inverted_index/mod.rs
@@ -5,7 +5,8 @@ use bitvec::vec::BitVec;
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::mmap_hashmap::{MmapHashMap, READ_ENTRY_OVERHEAD};
 use common::types::PointOffsetType;
-use memory::madvise::{AdviceSetting, clear_disk_cache};
+use memory::fadvise::clear_disk_cache;
+use memory::madvise::AdviceSetting;
 use memory::mmap_ops;
 use memory::mmap_type::{MmapBitSlice, MmapSlice};
 use mmap_postings::MmapPostings;
