+++ lib/segment/src/index/field_index/geo_index/mmap_geo_index.rs
@@ -6,7 +6,8 @@ use common::counter::hardware_counter::HardwareCounterCell;
 use common::types::PointOffsetType;
 use io::file_operations::{atomic_save_json, read_json};
 use memmap2::MmapMut;
-use memory::madvise::{AdviceSetting, clear_disk_cache};
+use memory::fadvise::clear_disk_cache;
+use memory::madvise::AdviceSetting;
 use memory::mmap_ops::{create_and_ensure_length, open_write_mmap};
 use memory::mmap_type::{MmapBitSlice, MmapSlice};
 use serde::{Deserialize, Serialize};
