+++ lib/segment/src/index/field_index/map_index/mmap_map_index.rs
@@ -13,7 +13,8 @@ use common::types::PointOffsetType;
 use io::file_operations::{atomic_save_json, read_json};
 use itertools::Itertools;
 use memmap2::MmapMut;
-use memory::madvise::{AdviceSetting, clear_disk_cache};
+use memory::fadvise::clear_disk_cache;
+use memory::madvise::AdviceSetting;
 use memory::mmap_ops::{self, create_and_ensure_length};
 use memory::mmap_type::MmapBitSlice;
 use serde::{Deserialize, Serialize};
