+++ lib/segment/src/vector_storage/chunked_mmap_vectors.rs
@@ -8,7 +8,8 @@ use common::maybe_uninit::maybe_uninit_fill_from;
 use io::file_operations::atomic_save_json;
 use memmap2::MmapMut;
 use memory::chunked_utils::{UniversalMmapChunk, chunk_name, create_chunk, read_mmaps};
-use memory::madvise::{Advice, AdviceSetting, clear_disk_cache};
+use memory::fadvise::clear_disk_cache;
+use memory::madvise::{Advice, AdviceSetting};
 use memory::mmap_ops::{create_and_ensure_length, open_write_mmap};
 use memory::mmap_type::MmapType;
 use num_traits::AsPrimitive;
