+++ lib/sparse/src/index/inverted_index/inverted_index_compressed_mmap.rs
@@ -10,7 +10,8 @@ use common::types::PointOffsetType;
 use io::file_operations::{atomic_save_json, read_json};
 use io::storage_version::StorageVersion;
 use memmap2::Mmap;
-use memory::madvise::{Advice, AdviceSetting, Madviseable, clear_disk_cache};
+use memory::fadvise::clear_disk_cache;
+use memory::madvise::{Advice, AdviceSetting, Madviseable};
 use memory::mmap_ops::{
     create_and_ensure_length, open_read_mmap, transmute_from_u8_to_slice, transmute_to_u8,
     transmute_to_u8_slice,
